Move to a pipeline design and split 'process' into stages
A huge refactoring was necessary to set up the pipeline. The 'input' and
'output' structures are now accessed from a 'FileRecord' that is being passed
throughout the pipeline.

The output behaviour to stderr and stdout has been overhauled: the diff now goes
to stderr while the index goes to stdout only if it is redirected.