Fix file management issues
'rmsrc' can not be used on multi-track files anymore as this would
prevent any track after #1 from being processed.

Covers are no longer processed if the audio file processing failed.

Previews show whether the source file is removed or processed in-place.

When a track fails for a multi-track files, the other tracks are still
processed.